+++ src/bin/cargo-list.rs
@@ -1,4 +1,4 @@
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use bunt::termcolor::ColorChoice;
 use cargo_list::Crates;
 use clap::{builder::TypedValueParser, Parser, ValueEnum};
@@ -6,6 +6,7 @@ use expanduser::expanduser;
 use indexmap::IndexSet;
 use is_terminal::IsTerminal;
 use rayon::prelude::*;
+use std::collections::BTreeMap;
 
 #[cfg(unix)]
 use pager::Pager;
@@ -43,7 +44,7 @@ enum Command {
 }
 
 #[derive(clap::Args)]
-#[command(version, long_about = None)]
+#[command(version, long_about = None, max_term_width = 80)]
 struct Cli {
     /// Output format
     #[arg(
@@ -69,13 +70,17 @@ struct Cli {
     #[arg(short, long)]
     outdated: bool,
 
+    /// Ignore version requirements
+    #[arg(short = 'I')]
+    ignore_req: bool,
+
     /// Consider a crate to be outdated if compiled with a Rust version
     /// different than the active toolchain
     #[arg(short = 'R')]
     outdated_rust: bool,
 
     /// Update outdated crates
-    #[arg(short, long, conflicts_with = "output_format")]
+    #[arg(short, long)]
     update: bool,
 
     /// Dry run
@@ -146,17 +151,43 @@ fn main() -> Result<()> {
     }
 
     let installed = Crates::from(&expanduser(&cli.config)?)?;
-    let all = installed.all();
-    let outdated = installed.outdated();
-    let outdated_rust = installed.outdated_rust();
+
+    let all = installed.crates();
+
+    let external = all
+        .par_iter()
+        .filter_map(|(name, c)| (c.kind == cargo_list::Kind::External).then_some((*name, *c)))
+        .collect::<BTreeMap<_, _>>();
+    let outdated = external
+        .par_iter()
+        .filter_map(|(name, c)| c.outdated.then_some((*name, *c)))
+        .collect::<BTreeMap<_, _>>();
+    let outdated_rust = external
+        .par_iter()
+        .filter_map(|(name, c)| c.outdated_rust.then_some((*name, *c)))
+        .collect::<BTreeMap<_, _>>();
+    let outdated_pinned = external
+        .par_iter()
+        .filter_map(|(name, c)| {
+            if !c.outdated && !c.outdated_rust && !c.newer.is_empty() {
+                Some((*name, *c))
+            } else {
+                None
+            }
+        })
+        .collect::<BTreeMap<_, _>>();
+
     let crates = if cli.outdated {
+        let mut crates = outdated.clone();
         if cli.outdated_rust {
-            &outdated_rust
-        } else {
-            &outdated
+            crates.append(&mut outdated_rust.clone());
         }
+        if cli.ignore_req {
+            crates.append(&mut outdated_pinned.clone());
+        }
+        crates
     } else {
-        &all
+        all.iter().map(|(name, c)| (*name, *c)).collect()
     };
 
     match cli.output_format {
@@ -183,7 +214,7 @@ fn main() -> Result<()> {
                     .collect::<Vec<_>>()
             };
 
-            let mut external = 0;
+            let mut ext = 0;
             for k in kinds {
                 bunt::println!("{$magenta+bold}# {:?}{/$}\n", k);
                 if k == cargo_list::Kind::External
@@ -191,38 +222,64 @@ fn main() -> Result<()> {
                     && (!cli.outdated || !outdated_rust.is_empty())
                 {
                     bunt::println!(
-                        "Active toolchain Rust version: {[bold]}\n",
-                        installed.current_rust
+                        "\
+                            Active toolchain:\n\n```text\n{}```\n\n\
+                            Active version: {[bold]}\n\
+                        ",
+                        installed.active_toolchain,
+                        installed.active_version,
                     );
                 }
                 let mut outdated = 0;
+                let mut update_pinned = 0;
                 let mut n = 0;
                 for c in all.values() {
                     if c.kind == k {
                         if k == cargo_list::Kind::External {
+                            let req = if let Some(req) = &c.version_req {
+                                if c.newer.is_empty() {
+                                    String::new()
+                                } else {
+                                    update_pinned += 1;
+                                    format!(" (Pinned: {req:?}; Available: {})", c.newer.join(", "))
+                                }
+                            } else {
+                                String::new()
+                            };
                             if c.outdated {
                                 bunt::println!(
-                                    "* {}: {[red]} => {}",
+                                    "* {}: {[red]} => {}{}",
                                     c.name,
                                     c.installed,
                                     c.available,
+                                    req,
                                 );
                                 n += 1;
                                 outdated += 1;
                             } else if cli.outdated_rust && c.outdated_rust {
                                 bunt::println!(
-                                    "* {}: {[green]} (Rust: {[red]})",
+                                    "* {}: {[green]} (Rust: {[red]}){}",
                                     c.name,
                                     c.installed,
                                     c.rust_version,
+                                    req,
+                                );
+                                n += 1;
+                                outdated += 1;
+                            } else if cli.ignore_req && !c.newer.is_empty() {
+                                bunt::println!(
+                                    "* {}: {[red]} => {}",
+                                    c.name,
+                                    c.installed,
+                                    &c.newer[0],
                                 );
                                 n += 1;
                                 outdated += 1;
                             } else if !cli.outdated {
-                                bunt::println!("* {}: {[green]}", c.name, c.installed);
+                                bunt::println!("* {}: {[green]}{}", c.name, c.installed, req);
                                 n += 1;
                             }
-                            external += 1;
+                            ext += 1;
                         } else if !cli.outdated {
                             bunt::println!("* {}: {[cyan]}", c.name, c.installed);
                             n += 1;
@@ -237,9 +294,9 @@ fn main() -> Result<()> {
                 if k == cargo_list::Kind::External {
                     if outdated == 0 {
                         bunt::println!(
-                            "{$green+italic}*All {} external crate{} are up-to-date!*{/$}\n",
-                            external,
-                            if external == 1 { "" } else { "s" },
+                            "{$green+bold}**All {} external crate{} are up-to-date!**{/$}\n",
+                            ext,
+                            if ext == 1 { "" } else { "s" },
                         );
                     } else {
                         bunt::println!(
@@ -248,35 +305,65 @@ fn main() -> Result<()> {
                             if outdated == 1 { "" } else { "s" },
                         );
                     }
+
+                    if !cli.ignore_req && update_pinned > 0 {
+                        bunt::println!(
+                            "\
+                                {$yellow+italic}*Consider updating {} pinned external crate{} via \
+                                `-I`.*{/$}\n\
+                            ",
+                            update_pinned,
+                            if update_pinned == 1 { "" } else { "s" },
+                        );
+                    }
                 }
             }
 
-            // Update crates
-            if cli.update
-                && (!outdated.is_empty() || (cli.outdated_rust && !outdated_rust.is_empty()))
-            {
-                for c in outdated.values().chain(outdated_rust.values()) {
-                    if c.kind == cargo_list::Kind::External {
+            // Update external crates
+            if cli.update {
+                let mut updates = outdated
+                    .iter()
+                    .map(|(name, c)| (*name, *c))
+                    .collect::<BTreeMap<_, _>>();
+                if cli.outdated_rust {
+                    updates.append(&mut outdated_rust.clone());
+                }
+                let mut update_pinned = 0;
+                if cli.ignore_req {
+                    updates.append(&mut outdated_pinned.clone());
+                    update_pinned += outdated_pinned.len();
+                }
+                if !updates.is_empty() {
+                    for (name, c) in &updates {
+                        let command =
+                            c.update_command(cli.ignore_req && outdated_pinned.contains_key(name));
                         if cli.dry_run {
-                            bunt::println!("```bash\n{$bold}{}{/$}", c.update_command().join(" "));
+                            bunt::println!("```bash\n{$bold}{}{/$}", command.join(" "));
                         } else {
-                            bunt::println!(
-                                "```text\n$ {$bold}{}{/$}",
-                                c.update_command().join(" ")
-                            );
-                            c.update();
+                            bunt::println!("```text\n$ {$bold}{}{/$}", command.join(" "));
+                            run(&command)?;
                         }
                         println!("```\n");
                     }
-                }
 
-                // Print summary
-                if !cli.dry_run {
-                    bunt::println!(
-                        "{$green+italic}*All {} external crate{} are up-to-date!*{/$}\n",
-                        external,
-                        if external == 1 { "" } else { "s" },
-                    );
+                    // Print summary
+                    if !cli.dry_run {
+                        bunt::println!(
+                            "{$green+italic}*All {} external crate{} are up-to-date!*{/$}\n",
+                            ext,
+                            if ext == 1 { "" } else { "s" },
+                        );
+                    }
+                    if !cli.ignore_req && update_pinned > 0 {
+                        bunt::println!(
+                            "\
+                                {$yellow+italic}*Consider updating {} pinned external crate{} via \
+                                `-I`.*{/$}\n\
+                            ",
+                            update_pinned,
+                            if update_pinned == 1 { "" } else { "s" },
+                        );
+                    }
                 }
             }
         }
@@ -284,13 +371,13 @@ fn main() -> Result<()> {
             #[cfg(unix)]
             Pager::with_pager("bat -pl json").setup();
 
-            println!("{}", serde_json::to_string(crates)?);
+            println!("{}", serde_json::to_string(&crates)?);
         }
         JsonPretty => {
             #[cfg(unix)]
             Pager::with_pager("bat -pl json").setup();
 
-            println!("{}", serde_json::to_string_pretty(crates)?);
+            println!("{}", serde_json::to_string_pretty(&crates)?);
         }
         Rust => {
             #[cfg(unix)]
@@ -307,4 +394,24 @@ fn main() -> Result<()> {
     }
 
     Ok(())
+}
+
+//--------------------------------------------------------------------------------------------------
+
+/**
+Run a command
+*/
+pub fn run(command: &[String]) -> Result<()> {
+    if std::process::Command::new(&command[0])
+        .args(&command[1..])
+        .spawn()
+        .unwrap()
+        .wait()
+        .unwrap()
+        .success()
+    {
+        Ok(())
+    } else {
+        Err(anyhow!("Command failed!"))
+    }
 }