+++ tests/integration.rs
@@ -1,5 +1,7 @@
 use cargo_list::Crates;
 use expanduser::expanduser;
+use rayon::prelude::*;
+use std::collections::BTreeMap;
 
 #[test]
 fn it_works() {
@@ -10,12 +12,16 @@ fn it_works() {
             if installed.is_empty() {
                 println!("No crates installed!");
             } else {
-                let outdated = installed.outdated();
+                let all = installed.crates();
+                let outdated = all
+                    .par_iter()
+                    .filter_map(|(&name, &c)| c.outdated.then_some((name, c)))
+                    .collect::<BTreeMap<_, _>>();
 
                 if outdated.is_empty() {
                     // List all crates in CSV
                     println!("Name,Installed");
-                    for (name, c) in &installed.all() {
+                    for (name, c) in &all {
                         println!("{name},{}", c.installed);
                     }
                 } else {